changed way to check time

+++ include/mandelbrot_alg.hpp
@@ -18,6 +18,6 @@ const int r_2_max = 10;
 const float ROI_X = -1.325f;  // Region Of Interest
 const float ROI_Y = 0;
 
-float RunMandelbrot_v1 (sf::Image* image, struct Params_t* cond, bool GraphicsFlag);
+double RunMandelbrot_v1 (sf::Image* image, struct Params_t* cond, bool GraphicsFlag);
 
 #endif // MANDELBROT_H
